requirements.txt
flask
python-dotenv
flask-cors
wtforms
flask_wtf
passlib
mysql-connector-python